# lib/api-zod/src/generated/types/createWorkInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateWorkInputState } from "./createWorkInputState";

export interface CreateWorkInput {
  name: string;
  shape: string;
  state: CreateWorkInputState;
}