# lib/api-zod/src/generated/types/updateWorkInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateWorkInputState } from "./updateWorkInputState";

export interface UpdateWorkInput {
  name: string;
  shape: string;
  state: UpdateWorkInputState;
}