# lib/api-zod/src/generated/types/work.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkState } from "./workState";

export interface Work {
  id: string;
  name: string;
  shape: string;
  state: WorkState;
  schemaVer: number;
  updatedAt: Date;
  createdAt: Date;
}